week_5/requirements.txt
click
boto3
python-dotenv
rich
